See extended description.
- Added DiscordUserProfileUser class (weird name because it's hard to fit in with everything else)
- Added UserProfile, Nitro and BoostingSince to DiscordProfile
- Added CertifiedModerator, Spammer (not a badge see comment in code), ActiveDeveloper badge to DiscordBadge enum
- Added GlobalName, BannerColor, AccentColor, AvatarDecoration to DiscordProfileUser
- Added Pronouns, Themecolors, PopoutAnimationParticleType, Emoji to DiscordUserProfileUser class
- Added AccountSettingsUpdate with alot of stuff moved from UserProfileUpdate
