+++ Anarchy/REST/User/Profile/UserProfileUpdate.cs
@@ -7,58 +7,6 @@ namespace Discord
     /// </summary>
     public class UserProfileUpdate
     {
-        private readonly DiscordParameter<string> NameProperty = new DiscordParameter<string>();
-        [JsonProperty("username")]
-        public string Username
-        {
-            get { return NameProperty; }
-            set { NameProperty.Value = value; }
-        }
-
-        public bool ShouldSerializeUsername()
-        {
-            return NameProperty.Set;
-        }
-
-        internal DiscordParameter<uint> DiscriminatorProperty = new DiscordParameter<uint>();
-        [JsonProperty("discriminator")]
-        public uint Discriminator
-        {
-            get { return DiscriminatorProperty; }
-            set { DiscriminatorProperty.Value = value; }
-        }
-
-        public bool ShouldSerializeDiscriminator()
-        {
-            return DiscriminatorProperty.Set;
-        }
-
-        private readonly DiscordParameter<string> EmailProperty = new DiscordParameter<string>();
-        [JsonProperty("email")]
-        public string Email
-        {
-            get { return EmailProperty; }
-            set { EmailProperty.Value = value; }
-        }
-
-        public bool ShouldSerializeEmail()
-        {
-            return EmailProperty.Set;
-        }
-
-        private readonly DiscordParameter<DiscordImage> AvatarProperty = new DiscordParameter<DiscordImage>();
-        [JsonProperty("avatar")]
-        public DiscordImage Avatar
-        {
-            get { return AvatarProperty; }
-            set { AvatarProperty.Value = value; }
-        }
-
-        public bool ShouldSerializeAvatar()
-        {
-            return AvatarProperty.Set;
-        }
-
         private readonly DiscordParameter<string> _bioParam = new DiscordParameter<string>();
         [JsonProperty("bio")]
         public string Biography
@@ -69,20 +17,14 @@ namespace Discord
 
         public bool ShouldSerializeBiography() => _bioParam.Set;
 
-        private readonly DiscordParameter<DiscordImage> _bannerParam = new DiscordParameter<DiscordImage>();
-        [JsonProperty("banner")]
-        public DiscordImage Banner
+        private readonly DiscordParameter<string> _proParam = new DiscordParameter<string>();
+        [JsonProperty("pronouns")]
+        public string Pronouns
         {
-            get { return _bannerParam; }
-            set { _bannerParam.Value = value; }
+            get { return _proParam; }
+            set { _proParam.Value = value; }
         }
 
-        public bool ShouldSerializeBanner() => _bannerParam.Set;
-
-        [JsonProperty("password")]
-        public string Password { get; set; }
-
-        [JsonProperty("new_password")]
-        public string NewPassword { get; set; }
+        public bool ShouldSerializePronouns() => _proParam.Set;
     }
 }
